FIX: standalone example

+++ examples/standalone/standalone.go
@@ -29,7 +29,7 @@ func main() {
 	}
 
 	numClients := 10
-	spawnRate := 10
+	spawnRate := float64(10)
 	globalBoomer = boomer.NewStandaloneBoomer(numClients, spawnRate)
 	globalBoomer.Run(task1)
 }
